pyproject.toml
[project]
name = "pygranso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "autograd>=1.8.0",
    "click>=8.3.0",
    "cma>=4.4.0",
    "gurobipy>=12.0.3",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy-stl>=3.2.0",
    "osqp[cu12]>=1.0.5",
    "pandas>=2.3.3",
    "pyomo>=6.9.5",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "torch-geometric>=2.7.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.2",
    "xarray>=2025.10.1",
    "yfinance>=0.2.66",
]
